+++ tf/modules/ecs/web_ec2/output.tf
@@ -1,9 +1,14 @@
 output "target_group_arn" {
   description = "ARN of ALB target group"
-  value       = aws_alb_target_group.service.arn
+  value       = module.target.target_group_arn
 }
 
 output "service_name" {
   description = "ECS Service name"
   value       = aws_ecs_service.service.name
+}
+
+output "service_arn" {
+  description = "ECS Service arn"
+  value       = aws_ecs_service.service.id
 }