+++ tf/modules/load-balancing/target/output.tf
@@ -1,0 +1,4 @@
+output "target_group_arn" {
+  description = "ARN of ALB target group"
+  value       = aws_alb_target_group.service.arn
+}