+++ changelog.md
@@ -60,6 +60,8 @@ Remove last usages of legacy interpolation syntax
 
 Add `min_size` and `max_size` to bastion module. This will allow setting to 0 to have a default 'off' bastion host.
 
-## 3.4 XXX
+## 3.4 2022-10
 
-Add `/ecs/ec2_capacity_provider` and `/ecs/web` modules.
+Add `/ecs/ec2_capacity_provider`, `/ecs/web_fargate` and `/ecs/web_ec2` modules.
+
+Added `/load-balancing/target` module use from above `/web_*` modules.